requirements.txt
pandas==2.3.3
python-jobspy==1.1.82
